backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.42.2
Deprecated==1.2.18
deprecation==2.1.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.11.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
gotrue==2.11.4
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
postgrest==0.19.3
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.4.1
requests==2.32.3
rich==13.9.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
Pillow
qrcode
reportlab